@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Calendar is a dictionary like Python object that can render itself as VCAL
 files according to rfc2445.
 
@@ -20,7 +19,6 @@ import pytz
 import dateutil.rrule, dateutil.tz
 from pytz.tzinfo import DstTzInfo
 
-from icalendar.compat import unicode_type
 
 
 ######################################
@@ -34,7 +32,7 @@ class ComponentFactory(CaselessDict):
     def __init__(self, *args, **kwargs):
         """Set keys to upper for initial dict.
         """
-        super(ComponentFactory, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self['VEVENT'] = Event
         self['VTODO'] = Todo
         self['VJOURNAL'] = Journal
@@ -60,7 +58,7 @@ _marker = []
 class Component(CaselessDict):
     """Component is the base object for calendar, Event and the other
     components defined in RFC 2445. normally you will not use this class
-    directy, but rather one of the subclasses.
+    directly, but rather one of the subclasses.
     """
 
     name = None         # should be defined in each component
@@ -79,7 +77,7 @@ class Component(CaselessDict):
     def __init__(self, *args, **kwargs):
         """Set keys to upper for initial dict.
         """
-        super(Component, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         # set parameters here for properties that use non-default values
         self.subcomponents = []  # Components can be nested.
         self.errors = []  # If we ignored exception(s) while
@@ -91,7 +89,7 @@ class Component(CaselessDict):
     #
     #    If the parser is too strict it might prevent parsing erroneous but
     #    otherwise compliant properties. So the parser is pretty lax, but it is
-    #    possible to test for non-complience by calling this method.
+    #    possible to test for non-compliance by calling this method.
     #    """
     #    return name in not_compliant
 
@@ -254,7 +252,7 @@ class Component(CaselessDict):
         return vals
 
     def set_inline(self, name, values, encode=1):
-        """Converts a list of values into comma seperated string and sets value
+        """Converts a list of values into comma separated string and sets value
         to that.
         """
         if encode:
@@ -335,7 +333,7 @@ class Component(CaselessDict):
                 component = stack[-1] if stack else None
                 if not component or not component.ignore_exceptions:
                     raise
-                component.errors.append((None, unicode_type(e)))
+                component.errors.append((None, str(e)))
                 continue
 
             uname = name.upper()
@@ -377,8 +375,7 @@ class Component(CaselessDict):
                 factory = types_factory.for_property(name)
                 component = stack[-1] if stack else None
                 if not component:
-                    raise ValueError('Property "{prop}" does not have '
-                                     'a parent component.'.format(prop=name))
+                    raise ValueError(f'Property "{name}" does not have a parent component.')
                 datetime_names = ('DTSTART', 'DTEND', 'RECURRENCE-ID', 'DUE',
                                   'FREEBUSY', 'RDATE', 'EXDATE')
                 try:
@@ -389,8 +386,7 @@ class Component(CaselessDict):
                 except ValueError as e:
                     if not component.ignore_exceptions:
                         raise
-                    component.errors.append((uname, unicode_type(e)))
-                    component.add(name, None, encode=0)
+                    component.errors.append((uname, str(e)))
                 else:
                     vals.params = params
                     component.add(name, vals, encode=0)
@@ -398,13 +394,21 @@ class Component(CaselessDict):
         if multiple:
             return comps
         if len(comps) > 1:
-            raise ValueError('Found multiple components where '
-                             'only one is allowed: {st!r}'.format(**locals()))
+            raise ValueError(cls._format_error(
+                'Found multiple components where only one is allowed', st))
         if len(comps) < 1:
-            raise ValueError('Found no components where '
-                             'exactly one is required: '
-                             '{st!r}'.format(**locals()))
+            raise ValueError(cls._format_error(
+                'Found no components where exactly one is required', st))
         return comps[0]
+
+    def _format_error(error_description, bad_input, elipsis='[...]'):
+        # there's three character more in the error, ie. ' ' x2 and a ':'
+        max_error_length = 100 - 3
+        if len(error_description) + len(bad_input) + len(elipsis) > max_error_length:
+            truncate_to = max_error_length - len(error_description) - len(elipsis)
+            return f'{error_description}: {bad_input[:truncate_to]} {elipsis}'
+        else:
+            return f'{error_description}: {bad_input}'
 
     def content_line(self, name, value, sorted=True):
         """Returns property as content line.
@@ -434,12 +438,8 @@ class Component(CaselessDict):
     def __repr__(self):
         """String representation of class with all of it's subcomponents.
         """
-        subs = ', '.join([str(it) for it in self.subcomponents])
-        return '%s(%s%s)' % (
-            self.name or type(self).__name__,
-            dict(self),
-            ', %s' % subs if subs else ''
-        )
+        subs = ', '.join(str(it) for it in self.subcomponents)
+        return f"{self.name or type(self).__name__}({dict(self)}{', ' + subs if subs else ''})"
 
 
 #######################################
@@ -612,12 +612,10 @@ class Timezone(Component):
                 tzname = component['TZNAME'].encode('ascii', 'replace')
                 tzname = self._make_unique_tzname(tzname, tznames)
             except KeyError:
-                tzname = '{0}_{1}_{2}_{3}'.format(
-                    zone,
-                    component['DTSTART'].to_ical().decode('utf-8'),
-                    component['TZOFFSETFROM'].to_ical(),  # for whatever reason this is str/unicode
-                    component['TZOFFSETTO'].to_ical(),  # for whatever reason this is str/unicode
-                )
+                # for whatever reason this is str/unicode
+                tzname = f"{zone}_{component['DTSTART'].to_ical().decode('utf-8')}_" + \
+                         f"{component['TZOFFSETFROM'].to_ical()}_" + \
+                         f"{component['TZOFFSETTO'].to_ical()}"
                 tzname = self._make_unique_tzname(tzname, tznames)
 
             dst[tzname], component_transitions = self._extract_offsets(
