@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """This module parses and generates contentlines as defined in RFC 2445
 (iCalendar), but will probably work for other MIME types with similar syntax.
 Eg. RFC 2426 (vCard)
@@ -6,9 +5,7 @@ Eg. RFC 2426 (vCard)
 It is stupid in the sense that it treats the content purely as strings. No type
 conversion is attempted.
 """
-from __future__ import unicode_literals
 
-from icalendar import compat
 from icalendar.caselessdict import CaselessDict
 from icalendar.parser_tools import DEFAULT_ENCODING
 from icalendar.parser_tools import SEQUENCE_TYPES
@@ -20,7 +17,7 @@ import re
 def escape_char(text):
     """Format value according to iCalendar TEXT escaping rules.
     """
-    assert isinstance(text, (compat.unicode_type, compat.bytes_type))
+    assert isinstance(text, (str, bytes))
     # NOTE: ORDER MATTERS!
     return text.replace(r'\N', '\n')\
                .replace('\\', '\\\\')\
@@ -31,16 +28,16 @@ def escape_char(text):
 
 
 def unescape_char(text):
-    assert isinstance(text, (compat.unicode_type, compat.bytes_type))
+    assert isinstance(text, (str, bytes))
     # NOTE: ORDER MATTERS!
-    if isinstance(text, compat.unicode_type):
+    if isinstance(text, str):
         return text.replace('\\N', '\\n')\
                    .replace('\r\n', '\n')\
                    .replace('\\n', '\n')\
                    .replace('\\,', ',')\
                    .replace('\\;', ';')\
                    .replace('\\\\', '\\')
-    elif isinstance(text, compat.bytes_type):
+    elif isinstance(text, bytes):
         return text.replace(b'\\N', b'\\n')\
                    .replace(b'\r\n', b'\n')\
                    .replace(b'\n', b'\n')\
@@ -53,12 +50,12 @@ def tzid_from_dt(dt):
     tzid = None
     if hasattr(dt.tzinfo, 'zone'):
         tzid = dt.tzinfo.zone  # pytz implementation
+    elif hasattr(dt.tzinfo, 'key'):
+        tzid = dt.tzinfo.key  # ZoneInfo implementation
     elif hasattr(dt.tzinfo, 'tzname'):
-        try:
-            tzid = dt.tzinfo.tzname(dt)  # dateutil implementation
-        except AttributeError:
-            # No tzid available
-            pass
+        # dateutil implementation, but this is broken
+        # See https://github.com/collective/icalendar/issues/333 for details
+        tzid = dt.tzinfo.tzname(dt)
     return tzid
 
 
@@ -71,7 +68,7 @@ def foldline(line, limit=75, fold_sep='\r\n '):
     immediately followed by a single linear white-space character (i.e.,
     SPACE or HTAB).
     """
-    assert isinstance(line, compat.unicode_type)
+    assert isinstance(line, str)
     assert '\n' not in line
 
     # Use a fast and simple variant for the common case that line is all ASCII.
@@ -146,7 +143,7 @@ def dquote(val):
     # so replace it with a single-quote character
     val = val.replace('"', "'")
     if QUOTABLE.search(val):
-        return '"%s"' % val
+        return f'"{val}"'
     return val
 
 
@@ -162,8 +159,7 @@ def q_split(st, sep=',', maxsplit=-1):
     length = len(st)
     inquote = 0
     splits = 0
-    for i in range(length):
-        ch = st[i]
+    for i, ch in enumerate(st):
         if ch == '"':
             inquote = not inquote
         if not inquote and ch == sep:
@@ -220,7 +216,7 @@ class Parameters(CaselessDict):
 
         for key, value in items:
             value = param_value(value)
-            if isinstance(value, compat.unicode_type):
+            if isinstance(value, str):
                 value = value.encode(DEFAULT_ENCODING)
             # CaselessDict keys are always unicode
             key = key.upper().encode(DEFAULT_ENCODING)
@@ -259,13 +255,13 @@ class Parameters(CaselessDict):
                     else:
                         result[key] = vals
             except ValueError as exc:
-                raise ValueError('%r is not a valid parameter string: %s'
-                                 % (param, exc))
+                raise ValueError(
+                    f'{param!r} is not a valid parameter string: {exc}')
         return result
 
 
 def escape_string(val):
-    # '%{:02X}'.format(i)
+    # f'{i:02X}'
     return val.replace(r'\,', '%2C').replace(r'\:', '%3A')\
               .replace(r'\;', '%3B').replace(r'\\', '%5C')
 
@@ -285,15 +281,15 @@ def unescape_list_or_string(val):
 #########################################
 # parsing and generation of content lines
 
-class Contentline(compat.unicode_type):
+class Contentline(str):
     """A content line is basically a string that can be folded and parsed into
     parts.
     """
     def __new__(cls, value, strict=False, encoding=DEFAULT_ENCODING):
         value = to_unicode(value, encoding=encoding)
         assert '\n' not in value, ('Content line can not contain unescaped '
-                                    'new line characters.')
-        self = super(Contentline, cls).__new__(cls, value)
+                                   'new line characters.')
+        self = super().__new__(cls, value)
         self.strict = strict
         return self
 
@@ -315,8 +311,8 @@ class Contentline(compat.unicode_type):
         values = to_unicode(values)
         if params:
             params = to_unicode(params.to_ical(sorted=sorted))
-            return cls('%s;%s:%s' % (name, params, values))
-        return cls('%s:%s' % (name, values))
+            return cls(f'{name};{params}:{values}')
+        return cls(f'{name}:{values}')
 
     def parts(self):
         """Split the content line up into (name, parameters, values) parts.
@@ -344,15 +340,13 @@ class Contentline(compat.unicode_type):
                                           strict=self.strict)
             params = Parameters(
                 (unescape_string(key), unescape_list_or_string(value))
-                for key, value in compat.iteritems(params)
+                for key, value in iter(params.items())
             )
             values = unescape_string(st[value_split + 1:])
             return (name, params, values)
         except ValueError as exc:
             raise ValueError(
-                "Content line could not be parsed into parts: '%s': %s"
-                % (self, exc)
-            )
+                f"Content line could not be parsed into parts: '{self}': {exc}")
 
     @classmethod
     def from_ical(cls, ical, strict=False):
@@ -374,6 +368,7 @@ class Contentlines(list):
     Then this should be efficient. for Huge files, an iterator should probably
     be used instead.
     """
+
     def to_ical(self):
         """Simply join self.
         """
