@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from datetime import datetime
 from icalendar.parser_tools import to_unicode
 from icalendar.prop import vDatetime
@@ -9,26 +8,26 @@ from string import digits
 import random
 
 
-class UIDGenerator(object):
+class UIDGenerator:
     """If you are too lazy to create real uid's.
 
     """
     chars = list(ascii_letters + digits)
 
-    def rnd_string(self, length=16):
+    @staticmethod
+    def rnd_string(length=16):
         """Generates a string with random characters of length.
         """
-        return ''.join([random.choice(self.chars) for _ in range(length)])
+        return ''.join([random.choice(UIDGenerator.chars) for _ in range(length)])
 
-    def uid(self, host_name='example.com', unique=''):
+    @staticmethod
+    def uid(host_name='example.com', unique=''):
         """Generates a unique id consisting of:
             datetime-uniquevalue@host.
         Like:
             20050105T225746Z-HKtJMqUgdO0jDUwm@example.com
         """
         host_name = to_unicode(host_name)
-        unique = unique or self.rnd_string()
+        unique = unique or UIDGenerator.rnd_string()
         today = to_unicode(vDatetime(datetime.today()).to_ical())
-        return vText('%s-%s@%s' % (today,
-                                   unique,
-                                   host_name))
+        return vText(f'{today}-{unique}@{host_name}')
