@@ -1,3 +1,2 @@
-# -*- coding: utf-8 -*-
-# we save all timezone with TZIDs unknow to the TZDB in here
+# we save all timezone with TZIDs unknown to the TZDB in here
 _timezone_cache = {}
